refactor(e2e): move the dummy-provider identity to the suite that uses it
UserIdentity and makeUserIdentity describe the backend's dummy-provider
account, not the msw mock: the mock has its own user record, and nothing
in @math3d/mock-api imported either. Every consumer is in
app-tests-e2e, and utils/api/auth.ts already re-exported the type.

Item 26 of the #1284 review.
