+++ packages/app-tests-e2e/src/tests/scenes-listing/deleting-scenes.test.ts
@@ -1,6 +1,7 @@
 import { test } from "@/fixtures/users";
 import { expect } from "@playwright/test";
-import { SceneBuilder, makeUserIdentity } from "@math3d/mock-api";
+import { SceneBuilder } from "@math3d/mock-api";
+import { makeUserIdentity } from "@/utils/api/auth";
 import type { Fixtures } from "@/fixtures/users";
 import AppPage from "@/utils/pages/AppPage";
